requirements.txt
Flask==3.0.0
sendgrid==6.12.4
gunicorn==21.2.0
python-dotenv==1.0.0
